+++ src/ResourceManager/DataMigration/Commands.DataMigration/Cmdlets/RemoveDataMigrationService.cs
@@ -1,13 +1,21 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="RemoveDataMigrationService.cs" company="Microsoft">
-//     Copyright (c) Microsoft Corporation.
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
 
+using System.Management.Automation;
 using Microsoft.Azure.Commands.DataMigration.Common;
 using Microsoft.Azure.Commands.DataMigration.Models;
 using Microsoft.Azure.Management.DataMigration.Models;
-using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
 {
@@ -46,7 +54,7 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
         [Parameter(
             Mandatory = true,
             ParameterSetName = ComponentNameParameterSet,
-            HelpMessage = "The name of the Data Migration Service.")]
+            HelpMessage = "The name of the Database Migration Service.")]
         [ValidateNotNullOrEmpty]
         [Alias("ServiceName")]
         public string Name { get; set; }