+++ src/ResourceManager/Media/ChangeLog.md
@@ -18,7 +18,11 @@
         - Additional information about change #1
 -->
 ## Current Release
+
+## Version 0.10.0
 * Set minimum dependency of module to PowerShell 5.0
+* Remove deprecated `Tags` alias from cmdlets
+    - `Set-AzureRmMediaService`
 
 ## Version 0.9.2
 * Updated to the latest version of the Azure ClientRuntime